backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
reportlab>=4.0.0
Pillow>=10.0.0
pymysql>=1.1.0
python-dateutil>=2.8.2